==> notebooks/02. cleaning_and_sentiment_analysis.ipynb <==
from pymongo import MongoClient
import re
from nltk.sentiment.vader import SentimentIntensityAnalyzer
import pandas as pd
from collections import defaultdict

#### Load tweets from mongoDB

client = MongoClient()
db = client.tweet_db
col = db.tweet_collection_full

col.find_one().keys()

def load_tweets_from_mongo_db(db, col):
    '''
    load tweets from mongo_db
    removes all retweets
    
    input: "db" (str) database name
    input: "col" (str) collection name
    
    returns a dictionary of tweets and twitter data
    '''
    
    all_tweets = defaultdict(list)
    for tweet in col.find({"full_text": {"$exists": True}}):
        if not tweet['retweeted'] and 'RT @' not in tweet['full_text'] and 'Retweeted' not in tweet['full_text']:
            all_tweets['tweet'].append(tweet['full_text'])
            all_tweets['date'].append(tweet['created_at'])
            all_tweets['handle'].append(tweet['user']['screen_name'])
            all_tweets['language'].append(tweet['lang'])
            all_tweets['id'].append(tweet['_id'])
    return all_tweets

tweets_dict = load_tweets_from_mongo_db(db = db, col = col)

df = pd.DataFrame(tweets_dict)
df.shape

df['date'] = pd.to_datetime(df['date'])

#### Drop non-English tweets

df1 = df.copy()

mask = df1['language'] != 'en'
df1 = df1.drop(df[mask].index, axis = 0)

df1.shape

#### Remove hypertext links and newlines from tweets

tweets = df1['tweet'].tolist()
remove = re.compile(r"http\S+|\n|")
cleaned_tweets = []
for tweet in tweets:
    tweet = remove.sub('', tweet).strip()
    cleaned_tweets.append(tweet)

df1['tweet'] = cleaned_tweets

#### Clean up the data by removing emojis, punctuation, twitter handles. This will help us peform sentiment analysis. Also I will see if there are any duplicate tweets in the data.

%run ../python_files/cleaning_helper.py # python module for cleaning tweets!

tweets_clean = []
tuples = []
tweets = df1['tweet'].tolist()
translator = str.maketrans('', '', remove_punctuaton) # punctuation remover
for tweet in tweets:
    for char in tweet:
    
        tweet = emojis.sub('', tweet)
    
    list_form = tweet.split() # turns the tweet into a list
    
    to_process = [x for x in list_form if not x.startswith("@")] # removes twitter handles
    
    string_form = " ".join(to_process) # back into a string
    
    set_form = set(string_form.translate(translator).strip().lower().split())
    
    tweets_clean.append(string_form.translate(translator).strip().lower())
    
    tuples.append(tuple(set_form)) # need to make it a tuple so it's hashable!

df1['tuples'] = tuples
df1['tweets_clean'] = tweets_clean

df1.shape

#### Drop duplicates

df2 = df1.copy()

df2 = df2.drop_duplicates(subset='tuples', keep="first") 
df2.shape

#### There were many duplicate tweets. This was because the same tweets were pulled multiple times or because some people were tweeting the exact same tweet.

#### Perform sentiment analysis
#### VADER is optimized for sentiment analysis in social media.
#### It understands slang, emoticons, and capitilization.

positive = []
negative = []
neutral = []
compound = []
sid = SentimentIntensityAnalyzer()
tweets = df2['tweets_clean'].tolist()
for tweet in tweets:
    ss = sid.polarity_scores(tweet)
    positive.append(ss['pos'])
    negative.append(ss['neg'])
    neutral.append(ss['neu'])
    compound.append(ss['compound'])

df2['positive'] = positive
df2['negative'] = negative
df2['neutral'] = neutral
df2['compound'] = compound
df2.shape

#### Create columns for each airline

handles = ['@united', '@Delta', '@SouthwestAir', '@AmericanAir']
airlines = ['United', 'Delta', 'Southwest', 'American']

for handle, airline in zip(handles, airlines):
    mask = df2['tweet'].str.lower().str.contains(handle.lower())
    df2.loc[mask, airline] = 1
    df2.loc[~mask, airline] = 0

df2.describe()

#### We are interested in the "compound" column. It is a measure of overall sentiment for a given tweet. Its range is from -1 for negative tweets, to +1 for positive tweets.

df2.loc[df2['compound'] == df2['compound'].max()].index

df2.loc[df2['compound'] == df2['compound'].min()].index

#### Highest and lowest sentiment scoring tweets

df2['tweet'][22502]

df2['tweet'][52782]

df2.to_pickle('../pickle_files/df1.p')
