==> notebooks/04. lemmatize.ipynb <==
import pandas as pd
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from nltk.tokenize import word_tokenize
import re

import spacy

%run ../python_files/cleaning_helper.py

df = pd.read_pickle('../pickle_files/df2.p')

df.shape

#### Lemmatize the tweets

nlp = spacy.load("en")

tweets = []
for i in range(len(df)):
    text = df.loc[i,'tweets_clean1']
    tokens = nlp(text)
    words = []
    for token in tokens:
        if token.is_space:
            continue
        elif token.text in remove_punctuaton: # from my cleaning module
            continue
        elif token.is_punct:
            continue
        elif token.is_stop:
            continue
        elif token.is_digit:
            continue
        elif len(token) <= 1:
            continue
        else:
            words.append(token.lemma_)
            
    tweet = ' '.join(words)
    tweets.append(tweet)

len(tweets)

row = np.random.choice(len(tweets))
tweets[row]

df['lemmatized_tweets'] = tweets

df.head()

df.to_pickle('../pickle_files/df3.p')
