==> notebooks/05. topic_modeling_lsa.ipynb <==
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
from sklearn.cluster import KMeans
from sklearn.decomposition import TruncatedSVD
from sklearn.pipeline import make_pipeline
from spacy.en.word_sets import STOP_WORDS
from sklearn.preprocessing import scale, Normalizer

plt.style.use('fivethirtyeight')

df = pd.read_pickle('../pickle_files/df3.p')

mask = df['compound'] <= -0.4
negative_rows = df.loc[mask] 
mask = df['compound'] >= 0.7
positive_rows = df.loc[mask]
mask = (df['compound'] > -0.4) & (df['compound'] < 0.7)
neutral_rows = df.loc[mask]

print('negative tweets:', len(negative_rows))
print('neutral tweets:', len(neutral_rows))
print('positive_tweets:', len(positive_rows))      

negative_rows.to_csv('../csvs/negative_rows.csv')
positive_rows.to_csv('../csvs/positive_rows.csv')

pd.read_csv('../csvs/negative_rows.csv', index_col='Unnamed: 0')

label = ['negative rows', 'neutral rows', 'positive rows']
num_rows = [len(negative_rows), len(neutral_rows), len(positive_rows)]
for l, n in zip(label, num_rows):
    print('{0}: {1}'.format(l, n))

df.columns

united_neg = negative_rows.loc[negative_rows['United'] == 1]
delta_neg = negative_rows.loc[negative_rows['Delta'] == 1]
sw_neg = negative_rows.loc[negative_rows['Southwest'] == 1]
aa_neg = negative_rows.loc[negative_rows['American'] == 1]

united_pos = positive_rows.loc[positive_rows['United'] == 1]
delta_pos = positive_rows.loc[positive_rows['Delta'] == 1]
sw_pos = positive_rows.loc[positive_rows['Southwest'] == 1]
aa_pos =  positive_rows.loc[positive_rows['American'] == 1]

airlines = ['United', 'Delta', 'Southwest', 'American']
neg_dfs = [united_neg, delta_neg, sw_neg, aa_neg]
pos_dfs = [united_pos, delta_pos, sw_pos, aa_pos]

stop = STOP_WORDS.union({'flight', 'amp', 'im', 'fly', 'flights'})
airline_stopwords = {'United': {'united'}, 
                  'Delta': {'delta', 'dea'},
                  'Southwest' : {'southwest'},
                  'American' : {'american', 'aa'}, 
                   }
all_stop = stop.union({'united', 'delta', 'dea', 'southwest', 'american', 
                       'aa', 'flight', 'amp', 'im', 'fly',
                      'good', 'thanks', 'great', 'thank', 'best', 
                       'like', 'awesome', 'today', 'airlines', 'better'})

#### SVD then clustering

def get_clusters_tfidf(airline, df, col, true_k = 5, n_components = 10, verbose = True, random_state = 24):
    vect = TfidfVectorizer(tokenizer = None,
                               max_df = 0.5, 
                               max_features = 200000, 
                               stop_words = all_stop,
                               use_idf = True, 
                               ngram_range=(1, 3),
                               min_df = 3)
    matrix = vect.fit_transform(df[col])
    grams = vect.get_feature_names()
    
    # do svd
    svd = TruncatedSVD(n_components = n_components, random_state = random_state)
    normalizer = Normalizer(copy=False)
    lsa = make_pipeline(svd, normalizer)
    reduced_matrix = lsa.fit_transform(matrix)
    
    # do kmeans
    km = KMeans(n_clusters = true_k, random_state = 24)
    km.fit(reduced_matrix)
    clusters = km.labels_.tolist()
    original_space_centroids = svd.inverse_transform(km.cluster_centers_)
    order_centroids = original_space_centroids.argsort()[:, ::-1]
    
    if verbose:
        original_space_centroids = svd.inverse_transform(km.cluster_centers_)
        order_centroids = original_space_centroids.argsort()[:, ::-1]
        print('\n', airline)
        for i in range(true_k):
            print("\nCluster %d:" % i)
            print(', '.join([grams[ind] for ind in order_centroids[i, :25]]))
    return km


kms_pos_tfidf_dict = {}
for airline, df in zip(airlines, pos_dfs):
    kms_pos_tfidf_dict[airline] = get_clusters_tfidf(airline, df, 'tweets_clean1', n_components = 10, 
                                                     true_k = 5, random_state = 42)
    df['cluster_num'] = kms_pos_tfidf_dict[airline].labels_ # append cluster number to df

kms_neg_tfidf_dict = {}
for airline, df in zip(airlines, neg_dfs):
    kms_neg_tfidf_dict[airline] = get_clusters_tfidf(airline, df, 'tweets_clean1', n_components = 10, 
                                                     true_k = 5, random_state = 42)
    df['cluster_num'] = kms_neg_tfidf_dict[airline].labels_ # append cluster number to df

kmeans_neg_american = kms_neg_tfidf_dict['American']
pd.Series(kmeans_neg_american.labels_).value_counts()

#### Map the clusters to the rows

rows_to_keep = ['Date', 'time', 'handle', 'tweet', 'compound', 'American', 'Delta', 'Southwest', 'United',
               'tweets_clean1', 'hashtags', 'cluster_num']
neg_dfs1 = []
for df in neg_dfs:
    df1= df.copy()
    df1 = df1[rows_to_keep]
    neg_dfs1.append(df1)

united_neg1 = neg_dfs1[0]
delta_neg1 = neg_dfs1[1]
sw_neg1 = neg_dfs1[2]
aa_neg1 = neg_dfs1[3]

for df in neg_dfs1:
    print(df.shape)

aa_neg1.to_csv('../csvs/aa_neg.csv')
delta_neg1.to_csv('../csvs/delta_neg.csv')
sw_neg1.to_csv('../csvs/sw_neg.csv')
united_neg1.to_csv('../csvs/united_neg.csv')

pos_dfs1 = []
for df in pos_dfs:
    df1= df.copy()
    df1 = df1[rows_to_keep]
    pos_dfs1.append(df1)

united_pos1 = pos_dfs1[0]
delta_pos1 = pos_dfs1[1]
sw_pos1 = pos_dfs1[2]
aa_pos1 = pos_dfs1[3]

for df in pos_dfs1:
    print(df.shape)

aa_pos1.to_csv('../csvs/aa_pos.csv')
delta_pos1.to_csv('../csvs/delta_pos.csv')
sw_pos1.to_csv('../csvs/sw_pos.csv')
united_pos1.to_csv('../csvs/united_pos.csv')
