==> notebooks/06. final_model.ipynb <==
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
from sklearn.decomposition import TruncatedSVD
from sklearn.cluster import KMeans
from sklearn.manifold import TSNE
from spacy.en.word_sets import STOP_WORDS
from sklearn.preprocessing import scale, Normalizer
from sklearn.pipeline import make_pipeline

plt.style.use('fivethirtyeight')

airlines = ['American', 'Delta', 'Southwest', 'United']
aa_neg = pd.read_csv('../csvs/aa_neg.csv', index_col = 'Unnamed: 0')
delta_neg = pd.read_csv('../csvs/delta_neg.csv', index_col = 'Unnamed: 0')
sw_neg = pd.read_csv('../csvs/sw_neg.csv', index_col = 'Unnamed: 0')
united_neg = pd.read_csv('../csvs/united_neg.csv', index_col = 'Unnamed: 0')
dfs_neg = [aa_neg, delta_neg, sw_neg, united_neg]

for df in dfs_neg:
    print(df.shape)

aa_pos = pd.read_csv('../csvs/aa_pos.csv', index_col = 'Unnamed: 0')
delta_pos = pd.read_csv('../csvs/delta_pos.csv', index_col = 'Unnamed: 0')
sw_pos = pd.read_csv('../csvs/sw_pos.csv', index_col = 'Unnamed: 0')
united_pos = pd.read_csv('../csvs/united_pos.csv', index_col = 'Unnamed: 0')
dfs_pos = [aa_pos, delta_pos, sw_pos, united_pos]

for df in dfs_pos:
    print(df.shape)

#### Create sentiment column

for df in dfs_pos:
    mask = df['compound'] <= -0.4
    df.loc[mask, 'negative_sent'] = 1.0
    mask = df['compound'] >= 0.7
    df.loc[mask, 'positive_sent'] = 1.0
    mask = (df['compound'] > -0.4) & (df['compound'] < 0.7)
    df.loc[mask, 'neutral_sent'] = 1.0
for df in dfs_neg:
    mask = df['compound'] <= -0.4
    df.loc[mask, 'negative_sent'] = 1.0
    mask = df['compound'] >= 0.7
    df.loc[mask, 'positive_sent'] = 1.0
    mask = (df['compound'] > -0.4) & (df['compound'] < 0.7)
    df.loc[mask, 'neutral_sent'] = 1.0

def label_sentiment(row):
    if row['neutral_sent'] == 1 :
        return 'Neutral'
    if row['positive_sent'] == 1 :
        return 'Positive'
    if row['negative_sent'] == 1 :
        return 'Negative'

for df in dfs_pos:
    df['sentiment'] = df.apply(label_sentiment, axis=1)
for df in dfs_neg:
    df['sentiment'] = df.apply(label_sentiment, axis=1)

cols_to_keep = ['Date', 'time', 'handle', 'tweet', 'tweets_clean1', 'hashtags', 'sentiment', 
                'United', 'Delta', 'Southwest', 'American']
for df in dfs_pos:
    df = df[cols_to_keep]
for df in dfs_neg:
    df = df[cols_to_keep]

for df in dfs_neg:
    print(df.shape)
for df in dfs_pos:
    print(df.shape)

df_whole = pd.concat([aa_neg, delta_neg, sw_neg, united_neg, aa_pos, delta_pos, sw_pos, united_pos])
df_whole['All Airlines'] = 1.0
df_whole.shape

unique = df_whole.loc[df_whole.index.value_counts() == 1].index.values
# sorted(unique)

df_whole.loc[unique].shape

df_whole1 = df_whole.loc[unique].copy()

df_whole1.rename(columns={'American': 'American Airlines', 
                         'Delta': 'Delta Airlines',
                         'Southwest': 'Southwest Airlines',
                         'United': 'United Airlines'}, inplace=True)

df_whole1.to_csv('../csvs/df_whole1.csv')

df_whole1 = pd.read_csv('../csvs/df_whole1.csv', index_col = 'Unnamed: 0')
df_whole1.shape

stop = STOP_WORDS.union({'flight', 'amp', 'im', 'fly', 'flights'})
airline_stopwords = {'United': {'united'}, 
                  'Delta': {'delta', 'dea'},
                  'Southwest' : {'southwest'},
                  'American' : {'american', 'aa'}, 
                   }
all_stop = stop.union({'united', 'delta', 'dea', 'southwest', 'american', 
                       'aa', 'flight', 'amp', 'im', 'fly',
                      'good', 'thanks', 'great', 'thank', 'best', 
                       'like', 'awesome', 'today', 'airlines', 'better'})

def get_clusters_tfidf(df, col, airline, true_k = 5, n_components = 20, verbose = True, random_state = 24):
#     df = df[df[airline] == 1] # convert the df to one with a single airline
    vect = TfidfVectorizer(tokenizer = None,
                               max_df = 0.5, 
                               max_features = 200000, 
                               stop_words = all_stop,
                               use_idf = True, 
                               ngram_range=(1, 3),
                               min_df = 3)
    matrix = vect.fit_transform(df[col])
    grams = vect.get_feature_names()
    
    # do svd
    svd = TruncatedSVD(n_components = n_components, random_state = random_state)
    normalizer = Normalizer(copy=False)
    lsa = make_pipeline(svd, normalizer)
    reduced_matrix = lsa.fit_transform(matrix)
    
    # do kmeans
    km = KMeans(n_clusters = true_k, random_state = 24)
    km.fit(reduced_matrix)
    clusters = km.labels_.tolist()
    original_space_centroids = svd.inverse_transform(km.cluster_centers_)
    order_centroids = original_space_centroids.argsort()[:, ::-1]
    
    if verbose:
        original_space_centroids = svd.inverse_transform(km.cluster_centers_)
        order_centroids = original_space_centroids.argsort()[:, ::-1]
        print('\n', airline)
        for i in range(true_k):
            print("\nCluster %d:" % i)
            print(', '.join([grams[ind] for ind in order_centroids[i, :25]]))
    return km

km_all = get_clusters_tfidf(df = df_whole, 
                 col = 'tweets_clean1', 
                 airline = 'All Airlines',
                 n_components = 50, 
                 true_k = 10, 
                 random_state = 42)

pd.Series(km_all.labels_).value_counts()

df_whole['cluster_num1'] = km_all.labels_

df_whole.to_csv('../csvs/df_final.csv')

df_whole.head()
